Set max datepicker date

+++ app/src/main/java/com/utn/nutricionista/models/Weight.kt
@@ -3,7 +3,7 @@ import java.time.LocalDate
 
 class Weight(var id:  String?,
              var uid : String? = null,
-             var peso: Double,
+             var peso: Float,
              var fecha : String)
 {
     fun date(): LocalDate {
